xnec2c-1.0b5.tar.gz import to git
This commit includes the following versions and change notes:

Version 0.1: First version with ability to draw a color-coded wire
frame Radiation Pattern (Near and Far Field), Graphs of various
Frequency dependent data (Gain, VSWR, Impedance etc) and the Structure
(Wires/Patches), including a color coded Current or Charge distribution.

Version 0.2: Incorporated some changes to the GUI (the Glade-generated
design) since after upgrading to GTK+ 2.8.9, the geometry of the windows
(the position and extend of buttons/entry widgets) changed a little. Also
fixed some bugs in the GUI code to handle unusual sequences of user
actions correctly.

Version 0.3: Added the ability to stop the frequency loop by clicking
on the frequency display spin buttons and to restart it by toggling the
"Gain" or "E/H" buttons in the Radiation Pattern window.

Version 0.4b: Added a NEC2 input file editor that makes the
edit/save/reload/execute cycle easier and quicker.

Version 0.5b:Fixed a bug that caused segmentation faults when only one
wire segment was present in the structure. This is not a case that will
normally exist but the seg fault had to be fixed.

Version 0.6b: Fixed a bug inherited from NEC2: If no geometry cards are
present (only a GE card) then there is division by zero in conect(). NEC2
seems to accept the lack of geometry cards, this is now an error condition
in xnec2c. Reduced the max value specification of the Capacitance
spin button in the Loading card editor, from 1.0e+12 to 1.0e+9 (pF)
since the higher value is beyond the range of the "long" type in 32 bit
systems. Modified the behavior of some card Editors to make them more
user friendly.

Version 0.7b: Fixed a bug in the co-ordinate translation (Move) editor
and edited the code in the GW card editor so that the wire diameter does
not change when the New button is clicked.

Version 0.8b: Fixed a serious bug that caused segmentation faults and
crashes in structures that have a wire in contact with a patch. The
subph() routine divides this patch into 3 or 4 sub-patches so the total
number of patches increases. There was no provision to reallocate and
initialize buffers used in rendering patches in the structure display
window. Other minor bugs were also fixed and the example NEC2 files were
also checked and edited when needed.

Version 0.9b: Fixed a few bugs mainly in the NEC2 editor code and also
changed the code of all individual Geometry and Command editors, so that
edited data are set in the main NEC2 editor's tree views when another
Geometry or Command card (tree view row) is selected for editing.

Implemented a work round around a serious bug which I could not trace,
since it now seems to be outside my code: When the zint() or fbar()
functions are called, they seem to corrupt xnec2c's memory allocation when
they return. This corruption manifests itself as NAN values appearing
in calculations so I modified these functions to return their computed
value via a pointer in the argument list. I also changed the function
declaration from complex long double to void as it seems this bug is
related to functions of the former type declaration, returning a value.

Version 1.0b: In this version xnec2c has been re-worked extensively
to make it multi-threading and to streamline its operation to some
extend. Many bugs created by these changes, and others that already
existed have been fixed, and the user interface has also changed
somewhat. The example input files have also been checked and some
mistakes in them have been corrected. This is now the first version 1.0
beta release for public testing.

Version 1.0b3: Added setlocale(LC_NUMERIC, "C"); to the main() function,
so that in locales where commas are used in decimal numbers xnec2c can
read data correctly (suggested by Joop Stakenborg). Some minor changes
to the user interface were also made to allow the NEC2 editor window to
fit in displays with shorter heights.

Version 1.0b4: After a bug report from Juha Vierinen I changed some
"sprintf" commands to "snprintf" to avoid buffer overruns.

Version 1.0b5: Following on the above changes, I revised all similar
situations in xnec2c's source and modified all "sprintf" commands to
"snprintf" just in case, as I could not replicate the bug so could not
test for other similar problems. I also fixed a bug in the "save" and
"save as" handler, to avoid false attempts by xnec2c to save structure
and radiation pattern/frequency plots pixmaps when a save of the NEC2
editor data failed for some reason.