xnec2c-1.2.tar.gz import to git
This commit includes the following versions and change notes:

Version 1.0: After several months with no bug reports or feature requests,
I am sticking my neck out and releasing xnec2c as version 1.0. This
version incorporates the last two feature request I received from users:
The "Cancel" button on card editors has been replaces with a "Delete Card"
button, which deletes the selected "card" (row in the NEC2 editor window).
A "Net Gain" menu item has been added to the View menu of the Frequency
Plots window to replace the second plotted quantity with the Net Gain
of the antenna.

Version 1.1: I spoke too soon! A segfault bug has been reported that
occurs when Maximum Admittance Matrix Asymmetry printing is requested in
the EX card. Although this is not supported in xnec2c, it still produces
a segmentation fault because the "ipnt" buffer is not allocated in the
netwk() function in network.c

Version 1.2: Made the page size of spin buttons 0 to make setting of
spin button values compatible with GTK 2.4