xnec2c-1.5.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 1.5: Changed the handling of command line arguments so that the
input file name may be specified without the use of the -i option. In
this case xnec2c will take the last argument to be the input file name,
but only if it has the .nec extension.  Got rid of some variables that
were set but not used, according to warnings given by gcc.  Applied a
patch supplied by Rik van Riel to allow the calculation of front to back
ratios when the antenna is modeled over ground.