xnec2c-2.3-beta.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 2.1: Patched a problem in the graph plotting routines that
caused xnec2c to crash with a division by zero fault. This can happen
when e.g. a single-wire structure is specified and a plot of gain vs
frequency is requested, for a polarization that is at 90 deg to the wire,
e.g. requesting gain plot for horizontal polarization for a vertical
wire. Since there is no radiation in the horizontal plane of a vertical
wire, the values of gain given to the plotting routines are all the same
so the vertical scale drawing routines crash xnec2c. Please note that
although xnec2c will now not crash as above, the plots will be jumbled
and meaningless.

Version 2.2-beta: I made extensive modifications to the source code to
silence a large number of warnings generated by the LLVM clang compiler
when used with the -Weverything option. These were mostly cases of
implicit conversions between variable types, like int to char or uint to
int etc. I have also updated the basic files of the GNU Autotools build
system, to be compatible with the current version of these tools at the
time of writing (February 2013).

Version 2.3-beta: Fixed a bug in the cell_edited_callback() function
that caused segmentation fault crashes when a cell in the input file
editor was edited directly by the user.