xnec2c-2.8.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 2.4: I submitted xnec2c (and nec2c) for scanning to the
Coverity source code audit website, which produced a list of no less
than 57 issues to be fixed! Most of them were not bugs that affected
calculations but possibly two, both in the Somnec code, one in function
rom1() and one in gshank() likely could have caused errors in the relevant
calculations. Unfortunately the FORTRAN source of Somnec (as well as that
of the NEC2) is very difficult to read so I cannot say with certainty
if this was so.

Version 2.5: After a bug report by Jean Collin, I made some changes to
the input file parser code in input.c, to properly identify tabs in the
input file.

Version 2.6: After a bug report by Lucjan SQ9VPA, I changed the case
of the signal handler that deals with SIGCHLD so that it doesn't cause
xnec2c to exit when the SIGCHLD signal does not originate from child
processes created by xnec2c. It appears that in some Linux installations
a SIGCHLD signal is sent to xnec2c even if it is not running forked,
e.g. the -j option is not used in the command line.

Version 2.7: After a bug report by Tim, WJ5Q, I fixed a bug that was
preventing the creation of an LD card of type 5 (LDTYP=5) when wire
conductivity (S/m) was specified in the GW card (Wire) editor. The same
bug was present in the GH (Helix) card editor and it was also fixed. I
added code so that the wire conductivity for GW, GA and GH cards can
be read from the relevant LD card and shown in the Conductivity (S/m)
spin button.

After all these changes I checked the xnec2c source code using the
Coverity Scan service and I fixed an out-of-bounds read error that was
reported by the scanner, in the plot_freqdat.c file.

Version 2.8: Fixed a bug in the GN card editor function which caused
xnec2c to save the GN card parameters in the .nec antenna description
file without clearing the number of radials to zero. This resulted in
the GN card editor window to open with confusing and incorrect defaults
for the radial screen, when the Reflection Coefficient Approximation
method was selected for specifying ground parameters.

Also fixed some bugs (missing variable initializations) in the Ground
Parameters calculation functions which resulted in incorrect Radiation
Pattern calculations. This would happen if a .nec file, with Perfect
Ground (iperf = 1) specified in the GN card, was opened after a .nec file,
with a Radial Ground Screen specified in the GN card, was processed.