xnec2c-3.3.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 2.9: After a request by the Debian maintainer of xnec2c, I added
a rudimentary man page he sent me and corrected some spelling errors
(arbitrary to arbitrary).

Version 3.0: I have modified the NEC2 file parser so that it can read
the file with both the '.' and ',' as decimal point characters. This
is necessary to make xnec2c compatible with different locales. I also
fixed a resource (memory) leak, reported by valgrind, in the xyz axes
drawing routine.

Version 3.1: I have made several changes to the code that renders wire
frame drawings and displays some data, like the projection parameters
(azimuth and elevation of the structure and the radiation pattern) and
gain values. This has resulted in a faster and smoother animation of
these drawings and reduced workload on the processor during dragging of
these images.  I have also reduced the default minimum size of xnec2c's
windows, as I had reports that they would not fit in the displays of
some laptops or monitors, with resolutions on the lower side.

Version 3.2: I made some changes to the Strlcat() function and its usage
in the xnec2c code, to improve safe handling of string concatenation
operations. Hopefully this has not broken the handling of various strings
in xnec2c! ;-)

Version 3.3: I fixed a bug that was in the code that reads the commands
from the NEC2 input file, where xnec2c would crash with a segmentation
fault when the FR card was after the RP card. Also modified the Strtod()
function to avoid freeing the end pointer before it was used, causing
problems.