xnec2c-3.5.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 3.4: I fixed a bug in the code that implements the modified GS
(scaling) card. In NEC2 the GS card scales all of the structure, but in
Xnec2c I modified the code so that it is possible to specify a range of
tag numbers to apply scaling to only. When a GS card follows a GX card,
selective scaling doesn't work correctly because a symmetric structure is
created. If you are using a GS card, it may be better to avoid specifying
a tag number range, just in case the bug fix is not complete.

Version 3.5: I modified the NEC2 Editor code so that when a Geometry
or Command editor window is opened, activating Apply or OK will save
data in the editor window to the treeview, even if the default data is
not edited.  I modified the NEC2 Editor code so that if a treeview row
is removed while the relevant editor window is open, then activating
Apply or OK will not attempt to save data to the treeview, since this
will cause an illegal memory access and will crash xnec2c.  Fixed a bug
in the Helix editor which caused incorrect calculation of the segment
length as a percentage of wavelength or of the number of segments/turn
that correspond to a given segment length as a percentage of wavelength.