xnec2c-3.9-beta.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 3.6-beta: I migrated Xnec2c to the latest GTK+-3.22 toolkit to
bring it up to date. A lot of changes were needed to both the GUI and
parts of the Xnec2c source code, mainly the use of GTK3's GUI Builder
for creating the User Interface and the revised drawing model of GDK for
presenting the Graphics displays. A new autogen.sh script has also been
incorporated in the source package and internationalization has been
migrated to the GNU GetText system. Because Xnec2c is very complicated
and thorough testing is difficult, I am releasing this version as a beta.

Version 3.6.1-beta: After a request by Don Walters W9DKI I modified the
xnec2c.glade file to reduce the default size of the drawing areas as
they would not fit in displays with reduced resolution. I also removed
some legacy code, left over from the migration of xnec2c to GTK3, which
was producing error messages from GTK3 during the dragging of structure
or radiation pattern displays. And I modified the Makefile.am file to
get "make install" to install program files and documentation to the
right places.

Version 3.7-beta: After some feature requests by users of xnec2c, I
implemented a number of changes to xnec2c so that it can save the state
of the GUI at exit and restore it on start up. To this effect xnec2c
now produces a configuration file xnec2c.config in the user's ~/.config
directory. In this file xnec2c saves the state (position, size, active
toggle buttons, projection parameters for structure and radiation pattern
displays, the state of the frequency loop etc) of the main windows,
e.g. structure, radiation pattern, frequency plots and NEC2 editor.
I have also added keyboard accelerators for a number of buttons and menu
items, more commonly used in the main windows of xnec2c, and a new "Zo"
NEC2 "card" to specify transmission line impedance.

Version 3.7.1-beta: Fixed a reported bug that prevented wire diameter
entries less than 0.01 in the Helix (GH) card editor. This was due to
specifying a minimum value of 0.01 in the xnec2c.glade file.

Version 3.8-beta: I edited some of the code so that when a NEC2 file
is saved, the Center x and y offsets (the change of position) of the
structure and radiation pattern are not changed. I also arranged for the
offset values to be saved in ~/.config/xnec2c.config when closing xnec2c,
and to be read back and restored when starting.

Version 3.9-beta: I have arranged for some user actions to be "nested",
mainly regarding the saving of NEC2 editor data followed by the opening
of another file or starting a new project. Some such functionality
already existed in previous versions but it is now more comprehensive.
Another useful change is fixing a bug in the original NEC2 code for
producing a helix (as specified by the GH "card"). This bug resulted a
division by zero if the helix length was specified as zero, possibly
intended to produce a spiral. The GH card can now be used to produce
both a helix or a spiral with no floating point exceptions.