xnec2c-4.1.1.tar.bz2 import to git
This commit includes the following versions and change notes:

Version 3.9: Since I have not received any bug reports for some time
after releasing version 3.9-b, and having fixed a bug that resulted in a
gtk_widget_destroy() command to be passed on an invalid widget pointer,
I am releasing xnec2c as version 3.9.

Version 4.0: I have changed the code that saves the state of the
View-Polarization menu items, as it would only correctly save the state
of these items in the Main (structure) window. I also changed the code
that handles the removal of "card" entries in the NEC2 editor, as it
was causing crashes of xnec2c under some user action combinations.

Version 4.0: I corrected some mistakes in the GUI interface code,
which handles the creation of windows and dialogs from the glade UI
description file. These mistakes were such that they did not effect
the operation of xnec2c but in any case I fixed them for the sake of
correct programming. Also changed the wire designer dialog to include
the Tapered Wire selection check-button in the dialog's frame and the
relevant code to deal with this better.

Version 4.1: I corrected some mistakes in the GUI interface code,
which handles the creation of windows and dialogs from the glade UI
description file. These mistakes were such that they did not effect
the operation of xnec2c but in any case I fixed them for the sake of
correct programming. Also changed the wire designer dialog to include
the Tapered Wire selection check-button in the dialog's frame and the
relevant code to deal with this better.

Version 4.1.1: I made a small change to the Radiation Pattern code
to allow the drawing of both horizontal an vertical "cuts" in the
radiation pattern, in the zx and zy vertical planes and the xy horizontal
plane. Also changed the accelerator keys for the Add, Remove, Clear,
Save As, Save and Revert buttons to be Modified by the Control button
as the original accelerator keys were being activated when writing in
the Comments Treeview.