Converted calls to Read/Write_Pipe() to use fixed data lengths and removed usleep() in Write_Pipe() Added a read_exact() and write_exact() call to replace the read/write() calls in Read/Write_Pipe() Fixed memory leak in Get_Freq_Data() if PRead_Pipe() fails Added mutexes in Frequency_Loop() and Open_Input_File() so they wont happen concurrently. Skip the inotify event if:   * the freqloop is running   * if CHILD is set   * or if any of forked_proc_data[job_num]->busy are set.
More detail:

In main.c's Open_Input_File() it sends open requests to child processes as follows:
	Write_Pipe( idx, fork_commands[INFILE], (ssize_t)lenc, TRUE );
	Write_Pipe( idx, rc_config.input_file,  (ssize_t)leni, TRUE );

A short time later it sends a request for frequency data in Frequency_Loop():
	/* Tell process to calculate freq dependent data */
	len = strlen( fork_commands[FRQDATA] );  // "frqdata"
	Write_Pipe( job_num, fork_commands[FRQDATA], (ssize_t)len, TRUE );

The child then reads this as a single request.  First it gets the
fork_commands[INFILE] string and tries to read the filename in
Child_Process's INFILE case:
	case INFILE: /* Read input file */
	  retval = Read_Pipe( num_child, rc_config.input_file, 80, FALSE );

But since "frqdata" was sent by the parent shortly after the filename,
the filename gets clobbered as "/tmp/test.necfrqdata".

The problem is in the attempt to mix variable-length and fixed-length data
transfers. and Write_Pipe() even hints at this because it has a usleep(10)
call and hopes to flush the pipe.  The kernel has no flushing mechanism
for fd's, so usleep() isn't a guarantee about flushing.  The behavior
could change from kernel version to kernel version, as well as per-host
configurations such as the NUMA configuration, CPU congestion, or even
the number of child processes forked by xnec2c.

Looking through the code it looks like the only variable-length
Read/Write_Pipe() call is for rc_config.input_file which is a static
length string.  I changed the appropriate lines above as follows:
	Write_Pipe( idx, rc_config.input_file, sizeof(rc_config.input_file), TRUE );
	...
	retval = Read_Pipe( num_child, rc_config.input_file, sizeof(rc_config.input_file), FALSE );

Added a read_exact() and write_exact() call to replace the read/write()
calls in Read/Write_Pipe() to guarantee the all of the data writes
(for example, because of EINTR).

Fixed memory leak cleanup and some places where buffers should be
initialized.  For example if PRead_Pipe() fails in Get_Freq_Data()
then it should free(buff) and return to try to continue gracefully.