* Added gtk context iterations to Stop() and Nec2_Save_Warn() because they block g_main by usleep'ing and will never draw the popup without it.
* Moved the check of whether cards are in correct order into the FR:
case because there was some kind of uninitialized datastructure issue
where it reported the error with min/max_mhz==0.  Tested and it still
works properly.

* Showed faulty NEC2 card name in the error and removed duplicated code.

* Added default return case to Read_Commands()

* Moved fr_plots initialization to its own function: fr_plots_init()

* Added additional correctness checks to fr_plot_sync_widths()

* Organized Set_Frequency_On_Click() so it doesn't return early by
creating variables to indicate whether to set fmhz or to redraw the
plot. If the plot should be redrawn, only redraw if global_lock can be
held to prevent fr_plot data from changing in the background (global_lock
prevents running the freqloop)