We were getting segfaults like this while sliding the gree frequency plot line during optimization: 	Program received signal SIGSEGV, Segmentation fault. 	0x0000000000480fd9 in Plot_Frequency_Data (cr=0x830b40) at plot_freqdata.c:1330 	1330	      gmax[idx] = rad_pattern[idx].gtot[mgidx] + Polarization_Factor(pol, idx, mgidx);
because mgidx is set like so before the call above:
	mgidx = rad_pattern[fstep].max_gain_idx[pol]

but inspecting mgidx shows it is a negative value:
	(gdb) p mgidx
	$1 = -172333112

This is happening because somewhere between the call
of Alloc_Rdpattern_Buffers() from input.c and the use of
rad_pattern[fstep].max_gain_idx[] in plot_freqdata.c (and elsewhere)
is not initialized in time by rdpat() in radiation.c so it faults in
calls like Plot_Frequency_Data() trying to access a negative or otherwise
invalid index.

We could try to figure out how to initialize it in time or add locks
to prevent use before initialization, but it is far easier to just
memset(*ptr, 0, req) inside of mem_alloc() to initialize all allocated
memory. Note that this does not fix memory that hasn't been initialized
by a call to mem_realloc() since mem_realloc() cannot portably know
what the original size was to zero the difference---so the caller must
initialize data after calls to mem_realloc() as necessary.