Converted all gtk_widget_queue_draw() calls to be wrapped by xnec2_widget_queue_draw().
This function provides us with a central place to control when a widget
is drawn based on the xnec2c flags.  For example, the radiation pattern
flashes the window grey if it is being moved while the frequncy loop is
running during optimization, thus we can disable drawing of rdpattern
in one place by checking isFlagSet(FREQ_LOOP_DONE).