Refactored Read/Save_Config()
Changed rc_config data types to be int's instead of uint8_t so they work
correctly with %d which is undefined on chars.

Updated Read/Save_Config to use a standard rc_config_vars_t structure
for defining the configuration.  This has the following benefits:
  * Read/Save_Config are simpler functions
  * Adding a configuration variable can be made in one place (rc_config_vars[])
  * Variables in the configuration file can be in any order
  * New variables will not break config compatability by requring a fixed order

Changed Create_Default_Config() so it verifies the xnec version and prints
a warning if versions differ, but this is no longer a fatal error because
the Read/Save_Config() functions are forwards and backwards compatible