Added option to Round X-axis Frequency Values in the way it did in v4.1.5
X-Axis frequency values in the Frequency Data Plots window default to
scaling based on the total frequency range and number of points which
deviates from the way Neoklis originally implemented the plot.  This makes
the best use of graph space, however, this new form can create axis values
that do not align nicely to whole numbers or common fractions of numbers
(like 13.5 vs 13.72).

The option "View->Round X Axis Frequency Values" will align the X Axis
labels to normalized values as originally designed in releases prior
to v4.1.5.  When this option is off, it will show one more frequency
signifigant figure displayed on the X Axis and scale the plot to use
the entire plot space.