Fixed regression caused by the -ffinite-math-only gcc optimization.
This gcc flag causes radiation patterns to draw incorrectly when the RP
card (radiation pattern card) when using a "Perfect Ground" in the GN
card at the same time that "Radial wire screen" is used as the ground
in the RP card.  Thank you Neoklis for investigating the examples that
produced the issue.

This means something in the radiation pattern code is generating either
+-Infinity or invalid floating point numbers (NaNs).  I'm not sure if
this should be fixed in NEC2, but it is pretty easy to fix by removing
the GCC flag and the performance changes is negligible.

This was introduced in v4.3 in commit b31c351c7641.