Updated examples with corrections from Neoklis.
The examples that had the problem of pattern display are incorrect,
specifying a radial screen with perfect ground. Also some had an incorrect
comment and some had the numerical data poorly tabulated, not important
but still preferable to have a regular format in the files.

If you specify non-perfect ground in GN, you can specify a radial
screen but only with the Reflection Coefficient option. With the
Sommerfeld-Norton method a radial screen is not acceptable. With a
non-perfect ground you would normally specify radials and they are
always assumed by NEC2 to be all connected at the origin of the x-y-z
coordinates.