Fixed deadlocks when running without pthreads and without forking.
When -P -j0 is specified, forking and the pthread frequency loop are
running synchronous for easier debugging and the locks for multi-threading
where deadlocking.  (Also there was a double-lock that should have been
a lock/unlock).

This is not a common use case for users because it is slow and choppy.