Initial framework to support NEC2 output files.
The original nec2c code read a .NEC input file and wrote a text
output file.  For xnec2c this was excluded because the text output was
spread across the calculation which made it difficult to integrate in
a meaningful way.

These work seeks to add that functionality to xnec2c based on the original
nec2c functionality in a way that it can operate independenly from the
frequency loop, and hopefully, without re-running the frequency loop to
generate the output.

This initial work does require calling New_Frequency() during each
iteration because crnt.cur[] must be populated for the current frequency.
This is because the existing xnec2c frequency loop does not save the crnt
data per-frequency like it does for other fields like impedance.  Future
work will address this by caching these datastructures per-frequency,
however at the cost of memory overhead, so it may become an option:

	To cache or not to cache, that is the question.

Certainly the user experience is better with caching, and most antenna
designs are small enough w.r.t. modern hardware that memory usage doesn't
matter too much.

The following outputs are currently implemented and the diffs are
consistent with the original nec2c output:

	COMMENTS
	STRUCTURE SPECIFICATION (does not display wires)
	SEGMENTATION DATA       (does not show "DATA CARD No" strings)
	FREQUENCY
	STRUCTURE IMPEDANCE LOADING
	ANTENNA ENVIRONMENT
	MATRIX TIMING           (hard-coded 0ms values for now)
	ANTENNA INPUT PARAMETERS
	CURRENTS AND LOCATION
	SEG TAG COORDINATES OF SEGM CENTER SEGM CURRENT (AMPS)
	POWER BUDGET