Added support for Touchstone 1- and 2-port files (.s1p, .s2p)
Touchstone files are used by professional RF software packages like
Microwave Office, Sonnet, and others.   Now xnec2c can export .s1p and
.s2p files to be used with those software packages.  S11 is the same as
return loss, reflection and .s1p files provide  single-port data.

For .s2p files gain is used as S21 and S12: we assume the antenna is
passive are passive so S21==S12.  S22 is a bit of a mystery, so we assume
that all S22 behavior is normalized into S11 and thus S22 is deminimus
and set it to -100 dB.  This may not be a correct assumption, so please
provide a suggestion here if you know a better way.

The S21 and S12 values in the .s2p come in two types:
  * Max Gain
  * Viewer Gain

Max gain is the maximum gain of the antenna, whereas, viewer gain is the
current gain of the antenna pointing toward the viewport of xnec2c.  For a
directional antenna, pointing the  main antenna lobe toward the xnec2c
viewport would give the same (or nearly the same) values as max gain.

Once you have your .s2p or .s1p files you can design a matching circuit
or other RF behavior in your favorite RF design software.

Related changes:
* The xnec2c "flags" are all most full so added a filechooser_t struct
  with callbacks and file extension types.

* Added meas_write_format() to write to a file descriptor

* Fixed prototype name in meas_format()