Changed default Save-As filenames to resemble the original .nec filename.
It used to be that menu items like "File->Save as gnuplot" would default
to filenames like "untitled.gplot".  This patch defaults them to the
NEC filename with the correct extension.  For example "antenna.nec"
would be saved as "antenna.gplot".  Same for .s1p, .s2p, and .png.

When you click Save As from the NEC editor it will still default to
"untitled.nec" to prevent an accidental overwrite of your existing
.nec file.