Disable EM calculation if the selected frequency step has already been calculated.
When the frequency loop runs it stores data into frequency-step-indexed
data structures.  You can change the selected frequency with the frequency
spinner widget or by clicking on the graph.  When you right-click on
a graph it selects the nearest calculated point.  Since that point
was already calculated with the frequency loop it skips the call to
New_Frequency() and saves some CPU time, especially on larger structures.

Left-clicking on a graph will (almost) always re-calculate the frequency
because it (probably) will not align to an existing calculated point
unless you manage to click within 1-Hz of a pre-calculated point, in
which case it will use the pre-calculated data.

Note that because the global structure `crnt` is not per-frequency it must
always re-calculate when "Draw Currents" or "Draw Charges" is enabled.
This is on the todo list to fix for later.

Additional cleanup:
	* Use meas_calc() in Display_Frequency_Data now that calc_data.members[]
      store data for the selected frequency, even if the selected frequency
	  was not originally calculated.

	* Renamed fstep to num_fsteps in _Plot_Frequency_Data() because usually
	  the variable `fstep` is an index into the frequency data, but in this case
	  it was being used as the number of steps available---which was confusing!

	* Changed frequency-error rounding from 1 kHz to 1 Hz so the cursor does
	  not jump around when selecting a frequency in the plot window.