Fixed complex S11 calculation for Touchstone file export
There was an error in the complex part of S11 calculation used to export
.s1p and .s2p Touchstone files (the real part was correct).  I thought the
complex column was the cimag(s11) but it turns out that it is cang(gamma).

I'm still trying to understand why it is cang(gamma) and not cang(s11),
maybe its just a design choice made by those who created the Touchstone
format?  Please comment if you know why.

Given the complex values:

	cgamma = (z_load-Zo) / (z_load+Zo);
	cs11 = 20*clog10( cgamma );

This first section is what we had initially, which is incorrect:

	creal(cs11);  # "Real" S11 column of the .s1p file (column 2)
	cimag(cs11);  # "Imag" S11 column of the .s1p file (column 3)

And this is the correct form:

	creal(cs11);  # "Real" S11 column of the .s1p file (column 2)
	cang(cgamma); # "Imag" S11 column of the .s1p file (column 3)

Thanks to Juan Pablo Garcia EA4CIV for discovering the issue and providing
the information needed to fix it!