Fixed loading ~/.xnec2c/xnec2c.conf with different locales.
If your locale changed then the xnec2c.conf file may have comma's
instead of periods for decimal points in floating-point config parameters
like radiation pattern position.  This can make the radiation pattern
"disappear" by going off the screen.

This patch solves that by adding a %n check to the scanf's and confirming
that all chars in the line are parsed.  If not all lines are parsed,
then it will retry with different locales in this order: "C", locale
LC_NUMERIC define, and finally an known comma-based decimal locale
"en_DK".  If none of these match then behavior is undefined, but this
will catch most if not all local mismatch issues.