Fixed crash caused memory allocation off-by-one error.
When the impedance_data.zreal and related arrays are allocated it does
not add +1 room for the "extra" selected frequency (the green line on
the plot).  Added +1 to accomodate this.  Thanks to Juan Pablo García,
EA4CIV for reporting the bug!

In addition, Draw_Graph() and Plot_Graph_Smith() were allocating
a points[] array for storing graph data, but they were based on the
number of frequency steps, not the number of plot points.  It turns out
these are equivalent because plots are always called to draw the number
of frequencies.  For clarity these functions were updated to use the
'nval' and 'nc' variable that is passed to them that indicate the number
of points being indexed.

Also added mem_obj_dump() while debugging memory objects.