Added directory permission hotfix for users affected by 4e62893b.
In commit 4e62893b the mkdir call used 755 instead of 0755 so permissions
were broken.  Since it only does this if ~/.xnec2c/ does not exists there
may be users out there with invalud permissions.  This commit hot-fixes
that for any affected users by doing a chmod (and respects the user's
umask setting).