mathlib: fixed duplicate BLAS detection, removed LMID support, simplified console output
The mathlib_t structure defines a linear algebra library that needs to
be dynamically linked.  Previously only one .so file was supported per
structure so if the same .so was symlinked with different matching names
then it would be listed multiple times in the menu.

Now multiple .so's can be specified comma-separated and it will search
to find the first matching library:

    {.type = MATHLIB_OPENBLAS, .lib = "libopenblas.so.0,libopenblas.so,libopenblas_serial.so.0",
        .name = "OpenBLAS+LAPACKe, Serial", .f_prefix = "LAPACKE_"},

LMID was only used to get the library directory path which is not critical
but has portability issues, so it was removed.

In addition, the console text was standardized as ".name:" so console
output is easier to read and understand when libraries are loaded.