Added support for GLIB v2.74
Last year g_idle_add_once() was implemented in xnec2c, and a year later the
Gnome team added that very function:
	https://thisweek.gnome.org/posts/2022/05/twig-45/

They may be implemented differently, but the only real difference between the
two versions in terms of functionality is the return value. This patch adds a
few #ifdef's that needed to deal with the change and fixes return values from
the xnec2c implementation of g_idle_add_once.

Thanks to the Debian maintainer Christoph Berg, DF7CB, for his initial work on
this! (f57e3b)