bs: Explicitly list all dist and install files
Use Automake variables to have Automake rules install files instead
of using a homebuilt "make install-data-local" recipe.

This also adds some files to the dist tarball which were
formerly forgotten (MAINTAINER README.md PACKAGING).

This also adds a long list of files (150+) spelled out explicitly
instead of a recursive copy operation just copying everything from
examples/.

This looks like a lot to maintain, but after this initial effort,
the changes should be only limited, when an antenna or two may be
added in the future.

We have also added an automatic consistency check which tells you
explicitly when that list is inconsistent what part of it is.

Advantage: If you add or remove any files in examples/, you
will need to reflect that change in Makefile.am which forces
you to think whether that change was intended or not, and
therefore this will catch unintended changes.

The recursive copy method just copied over everything, including
editor backup files, or anything else which happens to be in one
of the affected directories.