bs: Look for gmodule-2.0 via PKG_CHECK_MODULES
Look for gmodule-2.0 via PKG_CHECK_MODULES.

This removes the need for the AC_CHECK_LIB() check later.

Not sure why we need to explicitly link against gmodule-2.0
on macOS and not on Linux, but this certainly builds on both.

Fixes one of four from https://github.com/KJ7LNW/xnec2c/issues/16