Converted num_busy_procs>0 BUG to a pr_debug
Neoklis 5B4AZ discovered that the check of whether busy processes exist when
the loop is done may not indicate a bug.  Ultimately the software works fine.
This commit changes it to a pr_debug in case we need to troubleshoot in the
future, but at least it will not create a BUG output log that would alarm
users.