Differentiate between windows at 0,0 and never opened
It used to be that window positions were initialized to 0,0 if they have never
been used, but it is valid to have a window at 0,0.  Instead, we use -1,-1 as the
default/unopened window position.  This allows us to know if the window has been
opened for the first time so it can be aligned to not overlap with the main xnec2c
window.