Convert double-locking synchronization to use a condition variable.
I was using double-locking as a way to wait for another thread to finish, so
when it unlocks, then I get the lock and continue. Apparently that (maybe
hackish) synchronization method is not portable. So, g_cond_wait() is the
better way to do this.  This implementation is more clear, too.

This fixes bug #30

Thanks to @anoduck for testing and reporting the issue.