fix integer overflow in frequency selection
When the mouse clicks a frequency in the frequency graph plot, it rounds
the frequency selection. Here is some history:

  * In xnec2c v4.1.5, no rounding was performed and sub-Hz values could
    be selected.

  * In xnec2c v4.2, selections were rounded to 1 kHz using by
    multiplying the mHz value by 1000.0, casting to an int, and dividing
    it out.

  * In xnec2c v4.4.7, selections were rounded to 1 Hz using by
    multiplying the mHz value by 1e6, casting to an int, and dividing
    it out.

Since an int is (typically) a 32-bit datatype, its range is from
-2,147,483,647 to 2,147,483,647.  Thus, with Hz-resolution when the
frequency exceeds 2.147GHz, it wraps to a negative value.  Changing this
to a 64-bit int solves the problem.

This fixes #35

Thank you to @rklasen for reporting and troubleshooting the issue.