Fix off-by-one array index
Neoklis, 5B4AZ reports:

It surfaced when I used the -sanitize=undefined flag with clang. The bug
is in the fork.c file in this statements in Child_Process():

      case INFILE: /* Read input file */
        retval = Read_Pipe(
            num_child, rc_config.input_file, sizeof(rc_config.input_file), FALSE );
        rc_config.input_file[retval] = '\0';
        Child_Input_File();
        break;

The retval variable receives a value of 256, which is off-by-one since
rc_config.input_file has a size of 256. I fixed it by changing to:

    rc_config.input_file[retval-1] = '\0';